Do not always "block" keypresses when Notifications are shown
This was anoying me for a long time. When you have e.g. a zap error or
any other notification the first press exits the notification window and
to continue to zap you need to press twice a button. One for stop showing
the popup and another one to continue to zap.

Now by instantiate the notification window when the keypress is still
used in the InfoBar as input so you can quit the Notification and directly
continue to zap or even get a channellist, menu, numberzap, EPG etc.

This is only preformed on the latest Notification (or when there is only
one) and in case for that notification the close_on_any_key parameter is set.