When closing standby screen also remove record_event
When the RecordTimerEntry Quitmainloop is activated while the box is in
standby when coming out if standby the notification is removed, but there
is still a record_event left in Navigation that still may shut down the
box after a recording event ends which is set as an auto recording.

Now also only the notification and the record_event is removed as the
RecordTimerEntry Quitmainloop is really activated.