gdi: Remove unnecessary null checks, use accel for 8-bit picload
If "data_phys" is set, you can be sure that the accel manager is
up and running. No need to perform a check on that.
gAccel->getInstance() will never return NULL anyway...

Let picload use an accellerated surface if the format is 8-bit. This
speeds up the 8->32 rendering later on, and doesn't use that much
memory.

Fix compiler warnings on invalid gUnmanageSurface->gSurface conversions.