Add box-specific display support, in particular vuduo2
When a skin_display_(boxtype).xml exists, use that instead of the
default one. Create a bit of a vfd skin for the vuduo2 (which has
a 140x32 sized vfd).

Could be replaced with a "displayname" configure parameter, so
that a display skin is chosen based on machine features. Then the
--with-textlcd and similar options can be dropped.