Remove obsoleted isNextRecordTimerAfterEventActionAuto config
As far I can see this is not required anymore after the previous commit.
I could not find any plugin or so that did still ise it, but it cannot
ensure that if might be required by some plugin.

If so it could be possible we need to revert this commit.