Another finetuning regards deepstandby and record timers
When a recording with no "do nothing" afterwards is made followed by a
recording with "auto" afterwards then on the second recording the receiver will still go to deep standby. When we increase the standby counter as soon a button is pressed or a recording completes (button pressed is then also called)

Thanks for Dima73 for reporting this one.