Attempt to use the biggest local disk if /hdd/movie does not exist
Hopefully, this makes it possible to use a USB disk or similar.
The whole method of selecting a path has become way too complicated
now, and is a candidate for redesign (without ever using resolvePath
again)

This may lead to unexpectedly recording onto another disk in some
extreme situations though.

But at least it will simply record onto whatever happens to be
handy at the time, which is better than just giving up.