When editing alternatives keep focus on marked channels
When editing alternatives and you did browse thought the bouquets,
providers, satellites or all services the focus was on the currently
played service instead of the service you selected to change the
alternatives. Now the selected service will be searched for the first
alternative that is being available within a selected list.