Network: Allow hotplug for network adapters
This allows to insert (e.g.) a WiFi stick while E2 is running. The
adapter will then be visible in the network menu.
Activating the network still needs to be done manually though.
Future enhancement would be to re-configure and enable it when it
has already been configured (e.g. call activateInterface).