MovieSelection: Play next audio file regardless of selection bar pos
When playing audio, the next item would only play when the selection
bar was on it. This was to prevent unexpected movement of that bar.

Change the implementation so that it plays the next item without
moving the bar.

And change the MovieList to accommodate this. Also, do not reload the
whole list when the playback item changes, but only repaint the items
that were affected.