Do not stop a service when in standby the current time is unknown.
Allow one minute to try to obtain the current time from the currently
tuned service. When this is still not possible we consider that the
specific service could not be tuned. For startup to standby the box will
be put in standby via a notification and then afterwards a service is
started. this patch will also stop this started service after the box is
put in standby after one minute, which also allows the current time is
correctly set.