Fastscan: Start fastscan after 90 seconds in standby instead of 10 min
This allows somehowe the fastscan will also be performend for users who
put their settop box when into deepstandby as after a timer recording the box should boot instandby. I had to reduce this time to allow to perform the fastscan after the one minute tune to a channel for getting a correct transponder time
and before the tuning to the channel to record really starts.

Of course some timer recordings are required to allow the auto Fastscan to
be perfrormed. However I suggest most of us will use timer recordings.