For seek make OK button the same way as pause button when paused
Some remotes do not have a separate pause and play button (e.g. Xtrend), they have a common pause/play button. When this is the case the config config.seek.on_pause has no functionality. Usually by default this config is play. By adding the same decision to the OK button when seeking standard nothing is change. But from now on you are also able to config gop stap forward and return to previous speed for Xtrend boxes as well.

Later it can be consider to revert this partly with using remote control types in keymap.xml. Just as this has been done for some DMM remotes.