Standby: stop service as soon valid time is received by transponder
Previously we wait 60 seconds to try to get the time when the box is
started in standby. After these 60 seconds the service running in the
background is stopped even when no time is received (due to bad weather?).
Now the service in the background is stopped as soon a valid time is
received by the transponder. Usually this is earlier that the specific 60
seconds.