Revert "frontend: prefer DVB-T2 when both are available."
This reverts before commit

T2 transponders will always have System_DVB_T2.
If a transponder with System_DVB_T_T2 makes it into the lamedb,
this means no fallback to T2 has been made during the automatic scan,
so we are dealing with a T transponder.

So we should definitely not prefer a T2 tuner for
System_DVB_T_T2.