bouquets/alternatives: add search path for bouquet's and alternatives' files.
Search path:
- [enigma2_config]/.
- [enigma2_config]/bouquets
- [enigma2_config]/alternatives
In this order.

Rationale:
- less clutter in the enigma2 config directory (especially if
you have lots of alternatives)
- possibility to link or mount one of the subdirectories to a
shared disk e.g. on a NAS, so you can keep one set of
bouquets/alternatives for all receivers.

Drawback: new bouquets and alternatives will always be created in
the default location, so you'll have to move them manually, if you
want them in one of the non-default directories.