listboxservice: first try to solve anomaly when service list starts with number markers
When we use cable scan we also get a radio service, but the service numbers there do not
start with channel number 1. This leads to a side effect that you cannot correctly zap
through these radio channels via left/right. Only one direction works but it will skip
each time a channel. This patch at least solve the zap error, but it still has another
site effect as it makes the first number marker visiable on top which is not intended
and in addition that one could get stuck on indicated as selected. I suggest for now
this is cosmetic, but this new side effect should also be solved later... probably
the open source community might also help here...