Revert "ChannelSelection: Make it possible to hide userbouquets"
This reverts commit c3346022ebc246a714c661b0e041427ec40c0f48

As this patch was not really required. In OpenPLi hide bouquets was functioning
directly within the service list already a long time ago.
There was only an anomaly in case the first service in a list was hidden. As this
anomaly was solved this patch is really not required anymore.