Revert "picload: refactor filling surface data"
This reverts commit 169619b3cab6e6c046d5878ca62ecda364436916.

Since picload is used almost exclusively for showing pictures in the photo
viewer and similar plugins, it usually handles images that have roughly
the target size of the framebuffer, up to 1920x1080.
Just filling this with zeroes takes about the same amount of time as the
actual data copying. Filling the whole frame roughly halves performance.

The code isn't optimized for readability, but for speed and with good
reason.