Picload: support exif orientation tag
Implement Exif orientation support. Rotate/flip images so they
turn up right on the screen.  Default switched of. Picture
player needs to be adapted to enable auto orientation.  To do
this and to allow future changes for manual rotation/flipping
the resize functions are integrated in the decode() step.

Note: only implemented for jpeg as the Exif reader only
      supports exif in jpeg.

Sorry for the big change blob. The earlier revert of the
picload patches resulted in to many merge errors that I
had no energie for to fix.

Simple measurements show that decoding is now marginally
faster than before.

FIXME/TODO: Cannot remove resize functions yet as they
            are still needed by decodeThumb()