servicemp3: ignore processing of the same tags messages
Handling of every tags message creates evUpdatedInfo event,
even if we are dealing with the same tags messages

In gstreamer-1.6.0 this becomes problem, since we are
flooded with many of the same tags messages
many tags messages -> many evUpdateInfo events ->
 -> negative impact on enigma2 performance (spinners)

To fix this we check if merged stored tags message
(m_stream_tags) with new tags message is different
then stored tags message. If it's not then we don't
create evUpdatedInfo, since nothing is changed.