Revert " CA: Use descrambling demux when we descramble -> allocate demux from …
…start.

This solves an issue for services that are recognised as "pvr"
(fe == nullptr) and need to have a descrambling demux but don't get
one. At this point enigma assumes recordings and streams are always
descrambled (which is obviously not the case (anymore)).

In practice this was only an issue on a very limited set of receivers
where some of the demuxers are non-descrambling."