ServiceMP3: Optimize GStreamer start sequence
On start set PLAYING state and check return value of the change state.
On stop make sure that last change state was sucessfull, it prevents deadlock on fast channel zap.
When unpause is called make sure that we are in paused state before change state.
Move evStart event in NULL to READY state and update EPG on READY to PAUSED state.

Note: The get state on stop might block forever, so use 5 seconds timeout.
Note:  The m_state change to running when pipeline is prerolled, on ready to paused state.