servicemp3/record: unambiguous usage of User-Agent header
user-agent header could be currently set either by
config entry or by passing it in service reference
path string as an extra header.

Problem is that if we set it in service reference path
string, it's not working and user-agent is still
set from config entry or default value is used.

To avoid confusion of which option to use and to
allow to set different user-agent header for every
service, we allow to set it only in service
reference path string

http://forums.openpli.org/topic/39602-servicemp3-and-custom-user-agent-header/?view=findpost&p=517256