FrontendInfo: When we have more than x tuners indicate it condesed.
With option STRING in frontendInfo the frontend status is showed in a
string with color changes. Usually the unused tuners are displayed in
grey. But know when we have default more then 5 tuners the non used tuners
will not be tuned. You can change the default 5 by adding an integer as
option (e.g. STRING,10 shows max 10 tuner positions instead of 5)

This patch is more intended for the VU+Solo with FBC tuners capability.
note that previously we had with the default skin max 5 positions...