NimManager: DVB-C first otherwise Enigma2 might get confused
the logic in NimManager.py is a bit weird about the tuner allocation, if DVB-T shows up before DVB-C the tuner setup will be wrong after a reboot. While the whole tuner is in DVB-C mode (and also reports that correctly) enigma2 is internally set to dvb-t.
Customers reported that switching from DVB-S to DVB-C starts to fail because of that.

See http://forums.openpli.org/topic/40415-enigma2-external-tuner-bugfix/