eDebug: simplify ringbuffer/logOutput calls
The enigma ringbufer does not use the loglevel information passed
to it. So remove it and get rid of overloaded addToLogbuffer function.
Reduce logOutput signal to use char * instead of std::string, to
avoid expensive std::string usage.
Can savely do this as eDebug is the only user of the ringbuffer.