eDebug: quick fix to avoid weird output
vsnprintf() returns length of string as if there is an unlimited
buffer. As this return value is used for further indexing
into the print buffer it can easily result in writing or reading
outside the print buffer causing unpredictable result (maybe
even leading to enigma crashes.)

This patch is just a quick fix and will result in truncated
error strings.

A better solution needs to be implemented.