Revert "Enabled/disabled fallback tuner for recordings seperately"
 - it's confusing
 - it doesn't actually do what it suggests
 - there is no advantage really

When you use this option and set it to "no", the only difference
is that a recording that already couldn't be made in the first place,
now also won't be made using the fallback tuner, so it always fails.

This option works at the wrong level, at this point the user cannot
be informed so the only options are:

 - make the recording using the fallback tuner
 - don't make the recording at all (and thereby failing)

So this option is useless. It should be implemented at user-interface
level, i.e. in the timer conflict detection. If one doesn't disable
the conflict detection explicitly, fallback tuners won't be considered
for timer recordings anyway.