Solve issue when adding alternatives to previous last service
When you want to add alternatives to the previous last service within a
userbouquet all services were loaded with all marked. This issue at least
solves this topic.

Please note I'm not sure if this is the real solution. So it might be we
have to look to the real issue later on. But at least for now this issue
is solved