New log interface
Here is some experimental patch for debug levels.
It is a proof of concept as there are still many opinions
on how to do this. To get the discussions going (or finalised)
below follows a rationale:

To manage the amount of enigma log output, the log functions
have been changed and additional functionality has been added
to allow for level based output filtering.

The old eWarning(), eDebug() and eFatal() functions have been
consolidated into a single eDebugLow() function and reimplemented
using #define statements.

The eDebugLow() function implements additional functionality
to filter messages based on a debug level or a specific tag.

New eLog defines allow to explicitely set a loglevel or tag for a
debug message:

	eLog(int level, chr * fmt, ...)
	eLog(char *tag, chr * fmt, ...)
	eLogNoNewlineStart(int level, chr * fmt, ...)
	eLogNoNewlineStart(char *tag, chr * fmt, ...)
	eLogNoNewline(int level, chr * fmt, ...)
	eLogNoNewline(char *tag, chr * fmt, ...)

When a tag is given the message will only be printed when
the engima variable debugTag is set to the given tag. Otherwise
the message will be skipped. The tag is matched with the
strlen(debugTag) character so 'tag-levels' can be created by
using for instance eLog("mytag", ...), and eLog("mytag1", ...) and
eLog("mytag11", ...) etc. debugTag "mytag" will show all 3
eLogs(), while debugTag "mytag1" will only show the last 2.

When a level is given the message will only be printed when
level is smaller or equal to the engima variable debugLvl.
Otherwise the message will be skipped.

The default enigma debugTag is the empty string, but can be
changed using the Linux ENIGMA_DEBUG_TAG environment variable.

The default enigma debugLvl is 4 (and cannot be smaller than 0),
but can be changed using the Linux ENIGMA_DEBUG_LVL environment
variable.

(These environment variables are only read on enigma2 startup so
set them up before starting enigms2)

The eDebug()  functions will issue a level=4 message.
The eWarning() function will issue a level=2 message.
The eFatal()   function will issue a level=0 message.
All python print() calls will issue a level=4 message.

(Hence eFatal messages will always be printed!)

The new eDebugLow() function accepts a
    debuglevel or debugtag,
    flags,
    printf (varargs) constructed message.

The curently defined flags are
    _DBGFLG_NONEWLINE    do no add a newline at the end
    _DBGFLG_NOTIME       do not prefix with a timestamp
    _DBGFLG_FATAL        the message is for a FATAL error
                         and will case enigma to crashdump

Possible Future extentions:
- create python log functions allowing debug levels
- flag to print real time stamps instead of jiffy based.
  (note that relating engima log with kernel log will be much
   more difficult in that case)
- specify enigma debugLvl and debugTag via the enigma settings
  interface. Someone needs to setup some python code for that
  and then directly change change the enigma debugLvl and debugTag
  variables
  (tried using the eConfigManager but that seems to crash enigma ...)
- ...