frontend: Assume feSatellite when tuner is not multitype
In a loop-through environment there is a issue if the tuner with the main cable attached is not used as "master" tuner (preferred tuner).

Tuner A (0) - Unicable-LNB - loopcable to Tuner C - connected with C - E2 preferred tuner A
Tuner B (1) - Diseq switch - direct cable
Tuner C (2) - Unicable-LNB - loopcable to Tuner D - connected with D
Tuner D (3) - Unicable-LNB - cable to lnb

Preferred Tuner in E2 = Tuner A . All commands should only be send to lnb by tuner D which is connected to LNB.

When tuner is not multitype lets assume feSatellite as we used to.

More info: openmips/stbgui@6cfcc0f#commitcomment-18738154