[Language] Fix global text domain not set
This leads to untranslated texts in plugins when they rely on default translations from Enigma2.

If this really happened or not was unpredictable because some plugins already contained the above statement and fixed the whole system that way as a side effect ;) .
If NO such plugin was installed, default translations where broken for ALL plugins.